backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langgraph
chromadb
sentence-transformers
requests
pypdf
groq